+++ src/mappings/utils/ids.ts
@@ -66,3 +66,18 @@ export function getMsgStakeServiceId(msgStakeId: string, serviceId: string): str
 export function getStakeServiceId(entityStakedId: string, serviceId: string): string {
   return `${entityStakedId}-${serviceId}`;
 }
+
+// Returns the id of the relay for claim and proof
+export function getRelayId({
+applicationId,
+serviceId,
+sessionId,
+supplierId
+}: {
+  serviceId: string,
+  applicationId: string,
+  supplierId: string,
+  sessionId: string,
+}): string {
+  return `${supplierId}-${applicationId}-${serviceId}-${sessionId}`;
+}
